fix(P2BI-202): rename files
